move id2vec preprocess to cmd_entries module
